pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "drlearn"
dynamic = ["version"]
description = "A library for distributionally robust learning in the scikit-learn interface."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
    {name = "Ronak Mehta", email = "ronakdm@uw.edu"},
]
keywords = [
    "python",
    "machine learning",
    "distribtuionally robust optimization"
]
classifiers = [
    "Environment :: Console",
    "Programming Language :: Python",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Science/Research",
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "absl-py>=0.7.1",
  "numba>=0.58.1",
  "numpy>=1.21",
  "scikit-learn>=1.6.0",
]

[project.optional-dependencies]
examples=[
  "tqdm>=4.66",
  "jupyterlab>=4.1",
  "matplotlib>=3.5.0",
]

docs = [
    "sphinx>=6.0.0",
    "sphinx_book_theme>=1.0.1",
    "sphinxcontrib-katex",
    "sphinx-autodoc-typehints",
    "ipython>=8.8.0",  # 8.7.0 has ipython3 lexer error
    "myst-nb>=1.0.0",
    "matplotlib>=3.5.0",
    "sphinx-gallery>=0.14.0",
    "sphinx-collections>=0.0.1",
    "sphinx_contributors",
    "seaborn"
]

[tool.setuptools.packages.find]
include = ["README.md", "LICENSE"]
exclude = ["test_*.py"]
